+++ CHANGELOG.md
@@ -8,3 +8,4 @@
  - doc: add few instructions in readme and documentation files
  - doc: add few words about tags publication
  - add prefix based on branch
+ - delete local branches that are already removed in origin